requirements.txt
google-generativeai
PyGithub
semgrep